requirements.txt
discord.py
elasticsearch-async
sqlalchemy
textblob